Rework song parser to return a tuple of the parsed data and to take in a QByteArray instead of a filename.
